feat: added error handling. Input validation still needs improvement.
